@@ -17,39 +17,43 @@
   },
   "homepage": "https://github.com/monofuel/badMars-JS#readme",
   "dependencies": {
+    "@types/chai": "^4.0.4",
     "@types/core-decorators": "^0.10.32",
     "@types/dat-gui": "^0.6.3",
+    "@types/express": "^4.0.37",
     "@types/lodash": "^4.14.64",
     "@types/material-ui": "^0.18.0",
     "@types/prop-types": "^15.5.1",
     "@types/react": "^16.0.2",
-    "@types/react-bootstrap": "^0.0.52",
+    "@types/react-bootstrap": "^0.31.2",
     "@types/react-dom": "^15.5.3",
     "@types/react-router-dom": "^4.0.7",
     "@types/react-tap-event-plugin": "^0.0.30",
+    "@types/request": "^2.0.3",
+    "@types/rethinkdb": "^2.3.5",
     "@types/three": "^0.84.7",
     "@types/winston": "^2.3.2",
+    "@types/ws": "^3.0.2",
     "alea": "0.0.9",
     "axios": "^0.16.1",
-    "babel-eslint": "^7.2.3",
+    "babel-eslint": "^8.0.0",
     "babel-preset-env": "^1.4.0",
     "babel-preset-es2017": "^6.24.1",
     "bufferutil": "^3.0.0",
     "chai": "^4.0.2",
-    "core-decorators": "^0.19.0",
+    "core-decorators": "^0.20.0",
     "dat-gui": "^0.5.0",
     "ejs": "^2.5.2",
     "express": "^4.14.0",
     "fast-json-patch": "^2.0.4",
-    "flow-bin": "^0.49.1",
-    "flow-runtime": "^0.12.0",
+    "flow-bin": "^0.55.0",
+    "flow-runtime": "^0.14.0",
     "grpc": "^1.0.0",
     "hat": "0.0.3",
     "lodash": "^4.17.4",
     "mapstraction": "^1.0.1",
     "material-ui": "^0.19.0",
-    "node-context": "^0.7.0",
-    "parse-json": "^2.2.0",
+    "parse-json": "^3.0.0",
     "prop-types": "^15.5.10",
     "react": "^15.6.1",
     "react-bootstrap": "^0.31.2",
@@ -63,7 +67,7 @@
     "simplex-noise": "^2.2.0",
     "source-map-support": "^0.4.15",
     "stats.js": "^0.17.0",
-    "three": "^0.85.2",
+    "three": "^0.87.1",
     "three-collada-loader": "^0.0.1",
     "three-mtl-loader": "^1.0.1",
     "three-obj-loader": "^1.1.2",
@@ -73,24 +77,24 @@
     "utf-8-validate": "^3.0.1",
     "vorpal": "^1.11.4",
     "winston": "^2.3.1",
-    "ws": "^2.3.1"
+    "ws": "^3.2.0"
   },
   "devDependencies": {
-    "babel-plugin-react-transform": "^2.0.2",
+    "babel-plugin-react-transform": "^3.0.0",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-object-rest-spread": "^6.23.0",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-plugin-transform-strict-mode": "^6.18.0",
     "babel-preset-react": "^6.11.1",
     "babelify": "^7.3.0",
-    "browserify": "^13.1.0",
-    "eslint": "^3.14.0",
+    "browserify": "^14.4.0",
+    "eslint": "^4.7.2",
     "eslint-plugin-flowtype": "^2.30.0",
     "gulp": "^3.9.1",
-    "gulp-babel": "^6.1.2",
+    "gulp-babel": "^7.0.0",
     "gulp-concat": "^2.6.1",
     "gulp-sourcemaps": "^2.4.1",
-    "gulp-uglify": "^2.0.0",
+    "gulp-uglify": "^3.0.0",
     "livereactload": "^3.3.0",
     "tslint": "^5.7.0",
     "vinyl-buffer": "^1.0.0",
