@@ -6,9 +6,9 @@
 
 import env from '../config/env';
 import Client from '../net/client';
-import ws from 'ws';
-import MonoContext from '../util/monoContext';
-import querystring from 'querystring';
+import * as ws from 'ws';
+import Context from '../util/context';
+import * as querystring from 'querystring';
 
 import Logger from '../util/logger';
 import DB from '../db/db';
@@ -50,20 +50,19 @@ export default class Net {
 
 			logger.info(ctx, 'user connected', { username: user.name });
 
-			// $FlowFixMe: silly
-			info.req.user = user;
+			(info.req as any).user = user;
 
 			callback(true);
 		}
 
-		const wss = new WebSocketServer({ port: env.wsPort, verifyClient });
+		const wss = new WebSocketServer({ port: parseInt(env.wsPort), verifyClient });
 		wss.on('connection', (ws: ws) => {
 			new Client(ctx.create(), ws);
 		});
 		return Promise.resolve();
 	}
 
-	makeCtx(timeout?: number): MonoContext {
-		return new MonoContext({ timeout }, this.db, this.logger);
+	makeCtx(timeout?: number): Context {
+		return new Context({ timeout }, this.db, this.logger);
 	}
 }
