@@ -4,9 +4,9 @@
 //	website: japura.net/badmars
 //	Licensed under included modified BSD license
 
-import express from 'express';
-import path from 'path';
-import util from 'util';
+import * as express from 'express';
+import * as path from 'path';
+import * as util from 'util';
 
 import env from '../config/env';
 
@@ -14,7 +14,7 @@ import mainRoute from '../web/routes/main';
 import managementRoute from '../web/routes/management';
 import healthRoute from '../web/routes/health';
 
-import MonoContext from '../util/monoContext';
+import Context from '../util/context';
 
 import Logger from '../util/logger';
 import DB from '../db/db';
@@ -41,7 +41,7 @@ export default class WebService {
 		managementRoute(ctx.create(), app);
 		healthRoute(ctx.create(), app);
 
-		return new Promise((resolve: Function) => {
+		return new Promise<void>((resolve: Function) => {
 			const server = app.listen(env.wwwPort, () => {
 
 				const host = server.address().address;
@@ -53,8 +53,8 @@ export default class WebService {
 		});
 	}
 
-	makeCtx(timeout?: number): MonoContext {
-		return new MonoContext({ timeout }, this.db, this.logger);
+	makeCtx(timeout?: number): Context {
+		return new Context({ timeout }, this.db, this.logger);
 	}
 
 }
