@@ -4,10 +4,10 @@
 //	website: japura.net/badmars
 //	Licensed under included modified BSD license
 
-import r from 'rethinkdb';
-import _ from 'lodash';
+import * as r from 'rethinkdb';
+import * as _ from 'lodash';
 
-import MonoContext from '../util/monoContext';
+import Context from '../util/context';
 import Logger from '../util/logger';
 import { checkContext } from '../util/logger';
 
@@ -15,11 +15,13 @@ import DBChunk from './chunk';
 import DBUnit from './unit';
 import DBUnitStat from './unitStat';
 
+type ProfileKey = string;
+
 class DBCall {
 	profile: ProfileKey;
 	name: string;
-	ctx: MonoContext;
-	constructor(ctx: MonoContext, name: string) {
+	ctx: Context;
+	constructor(ctx: Context, name: string) {
 		this.ctx = ctx;
 		this.name = name;
 		this.profile = ctx.logger.startProfile(this.name);
@@ -34,19 +36,19 @@ class DBCall {
 	}
 }
 
-export async function createTable(conn: r.Connection, logger: Logger, tableName: string, primaryKey?: string): r.Table {
+export async function createTable(conn: r.Connection, logger: Logger, tableName: string, primaryKey?: string): Promise<r.Table> {
 	let results;
 	if (primaryKey) {
-		results = await r.tableList().contains(tableName).do((exists: boolean): any => {
-			return r.branch(exists, {
+		results = await (r as any).tableList().contains(tableName).do((exists: boolean): any => {
+			return r.branch(exists as any, {
 				table_created: 0
-			}, r.tableCreate(tableName, { primaryKey }));
+			}as any, (r as any).tableCreate(tableName, { primaryKey }));
 		}).run(conn);
 	} else {
-		results = await r.tableList().contains(tableName).do((exists: boolean): any => {
-			return r.branch(exists, {
+		results = await (r as any).tableList().contains(tableName).do((exists: boolean): any => {
+			return r.branch(exists as any, {
 				table_created: 0
-			}, r.tableCreate(tableName));
+			} as any, (r as any).tableCreate(tableName));
 		}).run(conn);
 	}
 	if (results.table_created) {
@@ -56,11 +58,11 @@ export async function createTable(conn: r.Connection, logger: Logger, tableName:
 }
 
 export async function createIndex(conn: r.Connection, logger: Logger, table: r.Table, name: string, multi?: boolean): Promise<void> {
-	const indexList = await table.indexList().run(conn);
+	const indexList: any = await table.indexList().run(conn);
 	if (multi) {
 		if (!indexList.includes(name)) {
 			logger.info(null, 'adding ' + name + ' index');
-			await table.indexCreate(name, { multi }).run(conn);
+			await table.indexCreate(name, { multi } as any).run(conn);
 		}
 	} else {
 		if (!indexList.includes(name)) {
@@ -71,17 +73,17 @@ export async function createIndex(conn: r.Connection, logger: Logger, table: r.T
 }
 
 export async function clearSpareIndices(conn: r.Connection, table: r.Table, validIndices: Array<string>): Promise<void> {
-	await table.indexWait();
+	await (table as any).indexWait();
 	const indexList: Array<string> = await table.indexList().run(conn);
 	const spareIndices = _.remove(indexList, (e: string): boolean => {
 		return !validIndices.includes(e);
 	});
-	for (const index: string of spareIndices) {
+	for (const index of spareIndices) {
 		await table.indexDrop(index).run(conn);
 	}
 }
 
-export function startDBCall(ctx: MonoContext, name: string): DBCall {
+export function startDBCall(ctx: Context, name: string): DBCall {
 	checkContext(ctx, `startDBCall ${name}`);
 	return new DBCall(ctx, name);
 }
