@@ -4,15 +4,17 @@
 //	website: japura.net/badmars
 //	Licensed under included modified BSD license
 
-import MonoContext from '../../util/monoContext';
+import Context from '../../util/context';
 import Client from '../client';
 import Unit from '../../unit/unit';
 import sleep from '../../util/sleep';
 import { sanitizeChunk, sanitizeUnit, sanitizePlanet } from '../../util/socketFilter';
 
-export default async function getMap(ctx: MonoContext, client: Client): Promise<void> {
+type ChunkHash = string;
+
+export default async function getMap(ctx: Context, client: Client): Promise<void> {
 	const units: Array<Unit> = await ctx.db.units[client.planet.name].listPlayersUnits(ctx, client.user.uuid);
-	const planet = sanitizePlanet(client.planet);
+	const planet: any = sanitizePlanet(client.planet);
 	planet.isSpawned = units.length !== 0;
 	client.send('map', { map: planet });
 
