@@ -4,11 +4,11 @@
 //	website: japura.net/badmars
 //	Licensed under included modified BSD license
 
-import MonoContext from '../../util/monoContext';
+import Context from '../../util/context';
 import Client from '../client';
 import User from '../../user/user';
 
-export default async function getPlayers(ctx: MonoContext, client: Client): Promise<void> {
+export default async function getPlayers(ctx: Context, client: Client): Promise<void> {
 	await ctx.db.user.listAllSanitizedUsers().then((users: Array<User>) => {
 		client.send('players', { players: users });
 	});
