@@ -4,11 +4,14 @@
 //	website: japura.net/badmars
 //	Licensed under included modified BSD license
 
-import hat from 'hat';
-import _ from 'lodash';
+const hat = require('hat');
+import * as _ from 'lodash';
 import { checkContext } from '../util/logger';
+import Context from '../util/context';
 
-class User {
+type TileHash = string;
+
+export default class User {
 	uuid: string; //filled by database
 	name: string;
 	color: string;
@@ -16,7 +19,7 @@ class User {
 
 	location: TileHash;
 
-	constructor(username: ? string, color: ? string) {
+	constructor(username: string = '', color: string = '') {
 		if(!username || !color) {
 			return; // load from DB
 		}
@@ -25,19 +28,15 @@ class User {
 		this.apiKey = hat();
 		this.color = color;
 	}
-	clone(other: Object) {
+	clone(other: any) {
 		for(const key in other) {
-			// $FlowFixMe: hiding this issue for now
-			this[key] = _.cloneDeep(other[key]);
+			(this as any)[key] = _.cloneDeep(other[key]);
 		}
 	}
 
-	async update(ctx: MonoContext, patch: Object): Promise<void> {
+	async update(ctx: Context, patch: Object): Promise<void> {
 		checkContext(ctx, 'user update');
 		Object.assign(this, patch);
 		await ctx.db.user.updateUser(this.name, patch);
 	}
-
 }
-
-module.exports = User;
